Made several changes, and started to add Spring Security.
For Dockerfile:
  Add clean to by gradle build line for consistency.
For build.gradle:
  Uncomment dependencies for Spring Security, and add new dependencies for spring-boot-testcontainers and junit-jupiter.
For checkstyle.xml:
  Comment out the line for "DesignForExtension"
For Hello.java:
  Add versioning to the path.
For QuadraticFormulaConsumerTest.java and QuadraticFormulaProducerTest.java:
  Start using testcontainers for the producer, and use newer annotations for both the producer and the consumer. Then, delete the no longer needed code, because now the annotations take care of this.
For the rest of the files, this is their initial commit.

With this commit, I am expecting to get an error message at this time.
Message = java.lang.AssertionError: Status expected:<200> but was:<401>
I shall continue to work on adding Spring Security at a later time.